// packages/panel-vendure/src/plugins/chasqui-plugin/customSurcharge/ui/generated-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
  /** A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  DateTime: any;
  /** The `Upload` scalar type represents a file upload. */
  Upload: any;
};
// ...
export type Node = {
  id: Scalars['ID'];
};
// ...
export type CustomSurcharge = Node & {
  __typename?: 'CustomSurcharge';
  id: Scalars['ID'];
  name: Maybe<Scalars['String']>;
  question: Maybe<Scalars['String']>;
  options: Maybe<Array<Maybe<CustomSurchargeOption>>>;
  enabled: Maybe<Scalars['Boolean']>;
};

export type CustomSurchargeAddInput = {
  enabled: Maybe<Scalars['Boolean']>;
  name: Scalars['String'];
  question: Scalars['String'];
  options: Maybe<Array<Maybe<CustomSurchargeOptionAddInput>>>;
};

export type CustomSurchargeOption = Node & {
  __typename?: 'CustomSurchargeOption';
  id: Scalars['ID'];
  listPrice: Maybe<Scalars['Int']>;
  sku: Maybe<Scalars['String']>;
  description: Maybe<Scalars['String']>;
};

export type CustomSurchargeOptionAddInput = {
  listPrice: Maybe<Scalars['Int']>;
  sku: Maybe<Scalars['String']>;
  description: Maybe<Scalars['String']>;
};

export type CustomSurchargeOptionUpdateInput = {
  id: Scalars['ID'];
  listPrice: Maybe<Scalars['Int']>;
  sku: Maybe<Scalars['String']>;
  description: Maybe<Scalars['String']>;
};

export type CustomSurchargeUpdateInput = {
  id: Scalars['ID'];
  enabled: Maybe<Scalars['Boolean']>;
  name: Scalars['String'];
  question: Scalars['String'];
  options: Maybe<Array<Maybe<CustomSurchargeOptionUpdateInput>>>;
};